requirements.txt
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-groq>=0.2.0
langchain-openai>=0.2.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
streamlit>=1.38.0
requests>=2.32.0
pandas>=2.2.0
tavily-python>=0.5.0
